requirements.txt
flask
flask-cors
pandas
numpy
python-dateutil
schedule
nltk
